Email confirmation post registration
User needs to verify before they can log in. Create global unique id
(GUID) every time a user is registered and sends it as an link to user
email. Add new actionresult in User and JobSearch controller and their
views. Made minor modifications in in user table and login. Made changes
to Apply actionresult to pass job id as parameter